Rewrite functions in async

+++ users/repository.py
@@ -1,7 +1,7 @@
 from typing import Type
 
 from fastapi import Depends
-from sqlalchemy.orm import Session
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from db import get_db
 from models import User
@@ -10,7 +10,7 @@ from query import get_user_by_id
 from schema import UserCreate, AdminUserCreate, UserUpdate
 
 
-def create_user(user: UserCreate, db: Session = Depends(get_db)) -> User:
+async def create_user(user: UserCreate, db: AsyncSession = Depends(get_db)) -> User:
     db_user = User(
         username=user.username,
         email=user.email,
@@ -20,14 +20,15 @@ def create_user(user: UserCreate, db: Session = Depends(get_db)) -> User:
         hashed_password=hash_password(user.password)
     )
 
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
+    async with db.begin():
+        db.add(db_user)
+        await db.commit()
+        await db.refresh(db_user)
 
     return db_user
 
 
-def create_admin_user(user: AdminUserCreate, db: Session = Depends(get_db)) -> User:
+async def create_admin_user(user: AdminUserCreate, db: AsyncSession = Depends(get_db)) -> User:
     db_user = User(
         username=user.username,
         email=user.email,
@@ -36,15 +37,19 @@ def create_admin_user(user: AdminUserCreate, db: Session = Depends(get_db)) -> U
         hashed_password=hash_password(user.password)
     )
 
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
+    async with db.begin():
+        db.add(db_user)
+        await db.commit()
+        await db.refresh(db_user)
 
     return db_user
 
 
-def update_user(user_id: int, user: UserUpdate, db: Session = Depends(get_db)) -> Type[User] | None:
-    db_user = get_user_by_id(user_id)
+async def update_user(user_id: int, user: UserUpdate, db: AsyncSession = Depends(get_db)) -> Type[User] | None:
+    db_user = await get_user_by_id(user_id, db)
+
+    if not db_user:
+        return None
 
     if user.first_name and db_user.first_name != user.first_name:
         db_user.first_name = user.first_name
@@ -55,13 +60,19 @@ def update_user(user_id: int, user: UserUpdate, db: Session = Depends(get_db)) -
     if user.phone_number and db_user.phone_number != user.phone_number:
         db_user.phone_number = user.phone_number
 
-    db.commit()
-    db.refresh(db_user)
+    async with db.begin():
+        await db.commit()
+        await db.refresh(db_user)
 
     return db_user
 
 
-def delete_user(user_id: int, db: Session = Depends(get_db)) -> None:
-    db_user = get_user_by_id(user_id)
-    db.delete(db_user)
-    db.commit()
+async def delete_user(user_id: int, db: AsyncSession = Depends(get_db)) -> None:
+    db_user = await get_user_by_id(user_id, db)
+
+    if not db_user:
+        return None
+
+    async with db.begin():
+        await db.delete(db_user)
+        await db.commit()
